% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser.R
\docType{data}
\name{ser}
\alias{ser}
\title{Data of sylvoecoregion}
\format{
An `sf` object in EPSG:2154 with 473 rows and 3 variables:
\describe{
  \item{codeser}{codeser}
  \item{NomSER}{NomSER}
  \item{geom}{geometry}
}
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?rubrique267}
}
\usage{
ser
}
\description{
TO-DO
}
\keyword{datasets}
