% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser_ar.R
\docType{data}
\name{ser_ar}
\alias{ser_ar}
\title{Data of sylvoecoregions of recent alluvial deposits}
\format{
An `sf` object in EPSG:2154 with 1048 rows and 3 variables:
\describe{
  \item{codeser}{codeser}
  \item{NomSER}{NomSER}
  \item{geom}{geometry}
}
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?rubrique267}
}
\usage{
ser_ar
}
\description{
TO-DO
}
\keyword{datasets}
