% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp.R
\docType{data}
\name{zp}
\alias{zp}
\title{Data of sylvoecoregions of recent alluvial deposits}
\format{
An `sf` object in EPSG:2154 with 1048 rows and 3 variables:
\describe{
  \item{AREA}{are}
  \item{PERIMETER}{perimeter}
  \item{ZPIFN250_}{ZPIFN250_}
  \item{ZPIFN250_I}{ZPIFN250_I}
  \item{ZP_RF}{ZP_RF}
  \item{LIBELLE}{libelle}
  \item{geom}{geometry}
}
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?rubrique267}
}
\usage{
zp
}
\description{
TO-DO
}
\keyword{datasets}
